refactor: add backend interface for poseidon builtin
